+++ psycopg/psycopg/cursor.py
@@ -9,7 +9,6 @@ from types import TracebackType
 from typing import Any, Generic, Iterable, Iterator, List
 from typing import Optional, NoReturn, Sequence, Tuple, Type, TypeVar
 from typing import overload, TYPE_CHECKING
-from warnings import warn
 from contextlib import contextmanager
 
 from . import pq
@@ -157,16 +156,17 @@ class BaseCursor(Generic[ConnectionType, Row]):
         Return `!True` if a new result is available, which will be the one
         methods `!fetch*()` will operate on.
         """
-        # Raise a warning if people is calling nextset() in pipeline mode
-        # after a sequence of execute() in pipeline mode. Pipeline accumulating
-        # execute() results in the cursor is an unintended difference w.r.t.
-        # non-pipeline mode.
+        # Python 3.1 would accumulate execute() results in the cursor, but it
+        # was an undesired difference with non-pipeline mode, so it was
+        # deprecated in 3.1.10; in 3.2 we moved to clobber previous execute()
+        # results and, because running multiple statements in the same
+        # execute() is not supported in pipeline mode, there is no reason to
+        # support nextset(). This error replaces the previous warning.
         if self._execmany_returning is None and self._conn._pipeline:
-            warn(
+            raise e.NotSupportedError(
                 "using nextset() in pipeline mode for several execute() is"
-                " deprecated and will be dropped in 3.2; please use different"
-                " cursors to receive more than one result",
-                DeprecationWarning,
+                " not supported. Please use several cursors to receive more"
+                " than one result"
             )
 
         if self._iresult < len(self._results) - 1: